[fixed] 1 of the zoning into 0,0 bugs [fixed] not seeing dropped objects in house bugs [fixed] wrong spawn
git-svn-id: http://svn2.assembla.com/svn/swgemu/core3/trunk@3299 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ MMOCoreORB/src/server/zone/managers/structure/StructureManagerImplementation.cpp
@@ -720,7 +720,7 @@ void StructureManagerImplementation::reportStructureStatus(CreatureObject* creat
 		ssmpool << dec << " (" << getTimeString(seconds) << ")";
 	}
 	*/
-	status->addMenuItem("@player_structure:maintenance_pool_prompt " + String::valueOf((int) floor(structure->getSurplusMaintenance())));
+	status->addMenuItem("@player_structure:maintenance_pool_prompt " + String::valueOf((int) floor((float)structure->getSurplusMaintenance())));
 	status->addMenuItem("@player_structure:maintenance_rate_prompt " + String::valueOf(structure->getBaseMaintenanceRate()) + " cr/hr");
 
 	if (structure->isInstallationObject()) {
@@ -774,7 +774,7 @@ void StructureManagerImplementation::promptManageMaintenance(CreatureObject* cre
 	//Get the most up to date maintenance count.
 	structure->updateStructureStatus();
 
-	int surplusMaintenance = (int) floor(structure->getSurplusMaintenance());
+	int surplusMaintenance = (int) floor((float)structure->getSurplusMaintenance());
 
 	ManagedReference<SuiTransferBox*> sui = new SuiTransferBox(creature, SuiWindowType::STRUCTURE_MANAGE_MAINTENANCE);
 	sui->setCallback(new StructureManageMaintenanceSuiCallback(server->getZoneServer()));
